requirements.txt
jiwer>=3.0.0
whisper-normalizer>=0.1.0
thefuzz>=0.20.0
python-Levenshtein>=0.20.0
numpy>=1.20.0
meeteval>=0.3.0
pyannote.metrics>=3.2.0
pyannote.core>=5.0.0
scipy>=1.7.0
requests>=2.28.0
python-dotenv>=0.19.0
